+++ web/src/components/plugin-store/hooks/use-plugin-store.ts
@@ -5,17 +5,20 @@ import {
   AnnotationState,
 } from "@/components/pdf-container/plugin-annotation-2"
 import type { DocumentManagerCapability } from "@/components/pdf-container/plugin-document-manager-2"
+import type { SearchCapability } from "@/components/pdf-container/plugin-search-2"
 import type { ScrollCapability } from "@/components/pdf-container/plugin-scroll-2"
 import type { SelectionCapability } from "@/components/pdf-container/plugin-selection-2"
 
 interface PluginStore {
   annoCapability: AnnotationCapability | null
   annoState: AnnotationState | null
+  searchCapability: SearchCapability | null
   selectCapability: SelectionCapability | null
   scrollCapability: ScrollCapability | null
   docManagerCapability: DocumentManagerCapability | null
   setAnnoCapability: (capability: AnnotationCapability | null) => void
   setAnnoState: (state: AnnotationState | null) => void
+  setSearchCapability: (capability: SearchCapability | null) => void
   setSelectCapability: (capability: SelectionCapability | null) => void
   setScrollCapability: (capability: ScrollCapability | null) => void
   setDocManagerCapability: (capability: DocumentManagerCapability | null) => void
@@ -25,11 +28,13 @@ interface PluginStore {
 const usePluginStore = create<PluginStore>((set) => ({
   annoCapability: null,
   annoState: null,
+  searchCapability: null,
   selectCapability: null,
   scrollCapability: null,
   docManagerCapability: null,
   setAnnoCapability: (annoCapability) => set({ annoCapability }),
   setAnnoState: (annoState) => set({ annoState }),
+  setSearchCapability: (searchCapability) => set({ searchCapability }),
   setSelectCapability: (selectCapability) => set({ selectCapability }),
   setScrollCapability: (scrollCapability) => set({ scrollCapability }),
   setDocManagerCapability: (docManagerCapability) => set({ docManagerCapability }),
@@ -41,6 +46,7 @@ export const usePluginCapabilities = () =>
   usePluginStore(
     useShallow((state: PluginStore) => ({
       annoCapability: state.annoCapability,
+      searchCapability: state.searchCapability,
       selectCapability: state.selectCapability,
       scrollCapability: state.scrollCapability,
       docManagerCapability: state.docManagerCapability,